refactor(widget): replace JLabel with TextPanel for better focus handling
- Switched from JLabel to TextPanel for the widget label
- Removed JBUI border styling
- Added focusable property to label
- Cleaned up unused imports and dependencies

I went to great lengths to fix the font jittering issue before, but it turned out not to be a code problem. Setting the font to "mono" in the code was useless; the user has to set the font directly in the IDE. In other words, we could have simply implemented the early TextWidget, but to minimize changes, we will not refactor it for now and will only replace the label with a TextPanel.